infra/aws/requirements.txt
attrs==20.2.0
aws-cdk.assets==1.68.0
aws-cdk.aws-applicationautoscaling==1.68.0
aws-cdk.aws-autoscaling-common==1.68.0
aws-cdk.aws-cloudwatch==1.68.0
aws-cdk.aws-codeguruprofiler==1.68.0
aws-cdk.aws-ec2==1.68.0
aws-cdk.aws-efs==1.68.0
aws-cdk.aws-events==1.68.0
aws-cdk.aws-iam==1.68.0
aws-cdk.aws-kms==1.68.0
aws-cdk.aws-lambda==1.68.0
aws-cdk.aws-logs==1.68.0
aws-cdk.aws-rds==1.68.0
aws-cdk.aws-s3==1.68.0
aws-cdk.aws-s3-assets==1.68.0
aws-cdk.aws-sam==1.68.0
aws-cdk.aws-secretsmanager==1.68.0
aws-cdk.aws-sqs==1.68.0
aws-cdk.aws-ssm==1.68.0
aws-cdk.cloud-assembly-schema==1.68.0
aws-cdk.core==1.68.0
aws-cdk.cx-api==1.68.0
aws-cdk.region-info==1.68.0
cattrs==1.0.0
constructs==3.0.20
jsii==1.13.0
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.3
